+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+- Allow specifying a subscription batch size
+
 ## [0.17.0] - 2018-03-22
 ### Added
 - Allow changing the event class using Event#with